chore: mv `DistinctSumAccumulator` to common
